requirements.txt
requests>=2.31.0
playwright>=1.40.0
Pillow>=10.0.0
pyautogui>=0.9.54
pywin32>=306
openpyxl>=3.1.0
tkcalendar>=1.6.1
pyinstaller>=6.0.0
